Refactor models, update database changelog
Refactored models in the application to make use of inheritance from a RootEntity class, reducing repeated code. The variables of the models have been changed from val to var to allow changes. Updated the database changelog to change the test data according to the new model structure, and updating the MemberController due to the removal of RequestCriteria.